requirements.txt
torch
torchvision
transformers
scikit-learn
pandas
numpy
matplotlib
tqdm
Pillow
streamlit
joblib